Conformance with ISO, level 3 warnings, and const rules.
ffscript.ypp was mostly hard to follow, some the code could be analyzed for const variables but most could not, there about 48 warnings about non-class const rules but are ignored for now. CMakeList.txt is not affected.

+++ src/parser/SemanticAnalyzer.cpp
@@ -954,7 +954,7 @@ void SemanticAnalyzer::caseExprCall(ASTExprCall& host, void* param)
 		// Count number of casts.
 		Function& function = **it;
 		int castCount = 0;
-		for (int i = 0; i < parameterTypes.size(); ++i)
+		for (size_t i = 0; i < parameterTypes.size(); ++i)
 		{
 			DataType const& from = getNaiveType(*parameterTypes[i], scope);
 			DataType const& to = getNaiveType(*function.paramTypes[i], scope);
